🩹 Fix: server component에서 데이터 가져오기
