+++ app/(home)/page.tsx
@@ -1,5 +1,20 @@
+import { getPostData } from '@/utils/getData';
 import Main from './Main';
 
-export default function HomePage() {
-  return <Main />;
+export interface Post {
+  id: string;
+  title: string;
+  content: string;
+}
+
+export default async function HomePage() {
+  const data = await getPostData();
+  const postItem: Post[] = data.map((post) => {
+    return {
+      id: post._id.toString(),
+      title: post.title,
+      content: post.content,
+    };
+  });
+  return <Main postItem={postItem} />;
 }
